# optimize N+1 query in retrieving CodeSchools with locations