--- a/app/serializers/code_school_serializer.rb
+++ b/app/serializers/code_school_serializer.rb
@@ -1,8 +1,4 @@
 class CodeSchoolSerializer < ActiveModel::Serializer
   attributes :name, :url, :logo, :full_time, :hardware_included, :has_online, :online_only, :mooc, :is_partner, :rep_name, :rep_email
   has_many :locations
-
-  def locations
-    object.locations.order(:state).order(:city)
-  end
 end