--- a/test/factories/code_schools.rb
+++ b/test/factories/code_schools.rb
@@ -12,5 +12,15 @@ FactoryGirl.define do
     is_partner false
     rep_name Faker::Name.name
     rep_email Faker::Internet.email
+
+    factory :code_school_with_locations do
+      transient do
+        locations_count { rand(1..15) }
+      end
+
+      after :create do |code_school, evaluator|
+        create_list(:location, evaluator.locations_count, code_school: code_school)
+      end
+    end
   end
 end