# Description of changes

In development, the query times went from (4 + number_of_schools)ms to 4ms.

* add a default sort order to the locations association on `CodeSchool`
* update `CodeSchoolsController` to include locations in the query result passed to the JSON rendering
* remove the custom `locations()` method from CodeSchoolSerializer to make the serializer rely on the association and default sort ordering already declared
* added a factory for creating a code school with 1-15 locations

# Issue Resolved

Partial fix for #395

# Showing Your Work

With the new `:code_school_with_locations` factory, I created 50 schools from the Rails console. I picked 50 based on the number of schools returned by the endpoint in the production site.

```
 > FactoryGirl.create_list(:code_school_with_locations, 50)
```

## Before this change:

238ms response time
* db: 50ms
* view: 232ms

<details>
<summary>Rails log for before</summary>

```
D, [2018-09-13T09:03:53.345273 #68534] DEBUG -- :   CodeSchool Load (1.9ms)  SELECT "code_schools".* FROM "code_schools" ORDER BY "code_schools"."name" ASC
D, [2018-09-13T09:03:53.390166 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 1]]
D, [2018-09-13T09:03:53.405458 #68534] DEBUG -- :   Location Load (1.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 2]]
D, [2018-09-13T09:03:53.408424 #68534] DEBUG -- :   Location Load (1.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 3]]
D, [2018-09-13T09:03:53.411751 #68534] DEBUG -- :   Location Load (1.1ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 4]]
D, [2018-09-13T09:03:53.413885 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 5]]
D, [2018-09-13T09:03:53.416255 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 6]]
D, [2018-09-13T09:03:53.418456 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 7]]
D, [2018-09-13T09:03:53.420373 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 8]]
D, [2018-09-13T09:03:53.422661 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 9]]
D, [2018-09-13T09:03:53.424990 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 10]]
D, [2018-09-13T09:03:53.427641 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 11]]
D, [2018-09-13T09:03:53.430295 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 12]]
D, [2018-09-13T09:03:53.432121 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 13]]
D, [2018-09-13T09:03:53.434300 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 14]]
D, [2018-09-13T09:03:53.436850 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 15]]
D, [2018-09-13T09:03:53.439364 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 16]]
D, [2018-09-13T09:03:53.442269 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 17]]
D, [2018-09-13T09:03:53.444533 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 18]]
D, [2018-09-13T09:03:53.446817 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 19]]
D, [2018-09-13T09:03:53.449239 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 20]]
D, [2018-09-13T09:03:53.451451 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 21]]
D, [2018-09-13T09:03:53.453364 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 22]]
D, [2018-09-13T09:03:53.455892 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 23]]
D, [2018-09-13T09:03:53.457855 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 24]]
D, [2018-09-13T09:03:53.460050 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 25]]
D, [2018-09-13T09:03:53.462613 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 26]]
D, [2018-09-13T09:03:53.465251 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 27]]
D, [2018-09-13T09:03:53.467947 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 28]]
D, [2018-09-13T09:03:53.470474 #68534] DEBUG -- :   Location Load (1.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 29]]
D, [2018-09-13T09:03:53.473131 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 30]]
D, [2018-09-13T09:03:53.475832 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 31]]
D, [2018-09-13T09:03:53.478969 #68534] DEBUG -- :   Location Load (1.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 32]]
D, [2018-09-13T09:03:53.481612 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 33]]
D, [2018-09-13T09:03:53.483544 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 34]]
D, [2018-09-13T09:03:53.485846 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 35]]
D, [2018-09-13T09:03:53.488553 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 36]]
D, [2018-09-13T09:03:53.491239 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 37]]
D, [2018-09-13T09:03:53.493970 #68534] DEBUG -- :   Location Load (1.1ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 38]]
D, [2018-09-13T09:03:53.497309 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 39]]
D, [2018-09-13T09:03:53.499375 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 40]]
D, [2018-09-13T09:03:53.501548 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 41]]
D, [2018-09-13T09:03:53.504538 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 42]]
D, [2018-09-13T09:03:53.507048 #68534] DEBUG -- :   Location Load (1.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 43]]
D, [2018-09-13T09:03:53.509662 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 44]]
D, [2018-09-13T09:03:53.511843 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 45]]
D, [2018-09-13T09:03:53.513802 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 46]]
D, [2018-09-13T09:03:53.516364 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 47]]
D, [2018-09-13T09:03:53.518418 #68534] DEBUG -- :   Location Load (0.7ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 48]]
D, [2018-09-13T09:03:53.520832 #68534] DEBUG -- :   Location Load (0.9ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 49]]
D, [2018-09-13T09:03:53.523430 #68534] DEBUG -- :   Location Load (0.8ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 50]]
D, [2018-09-13T09:03:53.524723 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 1]]
D, [2018-09-13T09:03:53.525689 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 2]]
D, [2018-09-13T09:03:53.527137 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 3]]
D, [2018-09-13T09:03:53.528841 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 4]]
D, [2018-09-13T09:03:53.529592 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 5]]
D, [2018-09-13T09:03:53.531023 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 6]]
D, [2018-09-13T09:03:53.532402 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 7]]
D, [2018-09-13T09:03:53.533242 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 8]]
D, [2018-09-13T09:03:53.534903 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 9]]
D, [2018-09-13T09:03:53.536704 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 10]]
D, [2018-09-13T09:03:53.538645 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 11]]
D, [2018-09-13T09:03:53.540291 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 12]]
D, [2018-09-13T09:03:53.541256 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 13]]
D, [2018-09-13T09:03:53.542746 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 14]]
D, [2018-09-13T09:03:53.544216 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 15]]
D, [2018-09-13T09:03:53.546299 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 16]]
D, [2018-09-13T09:03:53.548340 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 17]]
D, [2018-09-13T09:03:53.549449 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 18]]
D, [2018-09-13T09:03:53.550403 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 19]]
D, [2018-09-13T09:03:53.551746 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 20]]
D, [2018-09-13T09:03:53.553001 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 21]]
D, [2018-09-13T09:03:53.553963 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 22]]
D, [2018-09-13T09:03:53.555784 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 23]]
D, [2018-09-13T09:03:53.556679 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 24]]
D, [2018-09-13T09:03:53.557854 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 25]]
D, [2018-09-13T09:03:53.559871 #68534] DEBUG -- :   CACHE (0.1ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 26]]
D, [2018-09-13T09:03:53.562270 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 27]]
D, [2018-09-13T09:03:53.564073 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 28]]
D, [2018-09-13T09:03:53.565353 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 29]]
D, [2018-09-13T09:03:53.566666 #68534] DEBUG -- :   CACHE (0.0ms)  SELECT "locations".* FROM "locations" WHERE "locations"."code_school_id" = $1 ORDER BY "locations"."state" ASC, "locations"."city" ASC  [["code_school_id", 30]]
…